+++ ThirdPartyAdapters/applovin/applovin/src/test/kotlin/com/google/ads/mediation/applovin/AppLovinRewardedRendererTest.kt
@@ -80,7 +80,7 @@ class AppLovinRewardedRendererTest {
     verify(rewardedAdLoadCallback).onFailure(adErrorCaptor.capture())
     val capturedError = adErrorCaptor.firstValue
     assertThat(capturedError.code).isEqualTo(AppLovinErrorCodes.NO_FILL)
-    assertThat(capturedError.message.startsWith(ERROR_MSG_REASON_PREFIX)).isTrue()
+    assertThat(capturedError.message).startsWith(ERROR_MSG_REASON_PREFIX)
     assertThat(capturedError.domain).isEqualTo(APPLOVIN_SDK_ERROR_DOMAIN)
   }