+++ ThirdPartyAdapters/mytarget/mytarget/src/main/java/com/google/ads/mediation/mytarget/MyTargetSdkWrapper.kt
@@ -15,6 +15,7 @@
 package com.google.ads.mediation.mytarget
 
 import android.content.Context
+import com.my.target.ads.MyTargetView
 import com.my.target.ads.RewardedAd
 import com.my.target.common.MyTargetVersion
 
@@ -22,6 +23,8 @@ object MyTargetSdkWrapper {
   @JvmStatic
   val sdkVersion: String
     get() = MyTargetVersion.VERSION
+
+  @JvmStatic fun createBannerAd(context: Context): MyTargetView = MyTargetView(context)
 
   @JvmStatic
   fun createRewardedAd(slotId: Int, context: Context): RewardedAd = RewardedAd(slotId, context)