+++ ThirdPartyAdapters/pangle/pangle/src/test/kotlin/com/google/ads/mediation/pangle/renderer/PangleInterstitialAdTest.kt
@@ -129,7 +129,7 @@ class PangleInterstitialAdTest {
     verify(pagInterstitialRequest).setAdString(BID_RESPONSE)
     verify(pagInterstitialRequest).setExtraInfo(extraInfoCaptor.capture())
     val extraInfo = extraInfoCaptor.firstValue
-    assertThat(extraInfo.containsKey(ADMOB_WATERMARK_KEY)).isTrue()
+    assertThat(extraInfo).containsKey(ADMOB_WATERMARK_KEY)
     assertThat(extraInfo[ADMOB_WATERMARK_KEY]).isEqualTo(WATERMARK)
     verify(pangleSdkWrapper)
       .loadInterstitialAd(eq(PLACEMENT_ID_VALUE), eq(pagInterstitialRequest), any())