+++ ThirdPartyAdapters/pangle/pangle/src/test/kotlin/com/google/ads/mediation/pangle/renderer/PangleNativeAdTest.kt
@@ -170,7 +170,7 @@ class PangleNativeAdTest {
     verify(pagNativeRequest).setAdString(BID_RESPONSE)
     verify(pagNativeRequest).setExtraInfo(extraInfoCaptor.capture())
     val extraInfo = extraInfoCaptor.firstValue
-    assertThat(extraInfo.containsKey(ADMOB_WATERMARK_KEY)).isTrue()
+    assertThat(extraInfo).containsKey(ADMOB_WATERMARK_KEY)
     assertThat(extraInfo[ADMOB_WATERMARK_KEY]).isEqualTo(WATERMARK)
     verify(pangleSdkWrapper).loadNativeAd(eq(PLACEMENT_ID_VALUE), eq(pagNativeRequest), any())
   }