+++ ThirdPartyAdapters/bigo/CHANGELOG.md
@@ -1,7 +1,14 @@
 ## Bigo Android Mediation Adapter Changelog
 
-#### Next Version
+#### Version 5.8.2.0 (In progress)
+
+#### Version 5.8.0.0
 - Added property to build the adapter with GMA Next-Gen SDK dependency.
+- Verified compatibility with Bigo Ads SDK 5.8.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.1.0.
+- Bigo Ads SDK version 5.8.0.
 
 #### Version 5.7.1.0
 - Verified compatibility with Bigo Ads SDK 5.7.1.