+++ ThirdPartyAdapters/imobile/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## i-mobile Android Mediation Adapter Changelog
 
-#### Next Version
+#### Version 2.3.2.3
 - Filter age-restricted requests for i-mobile SDK because it cannot receive age-restricted signals.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.1.0.
+- IMobile SDK version 2.3.2.
 
 #### Version 2.3.2.2
 - Added property to build the adapter with GMA Next-Gen SDK dependency.