+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 Built and tested with:
 - Google Mobile Ads SDK version 25.1.0.
+- Google Mobile Ads Next-Gen SDK version 0.25.0-beta01.
 - InMobi Kotlin SDK version 11.2.0.
 
 #### Version 11.1.1.1