+++ ThirdPartyAdapters/maio/CHANGELOG.md
@@ -1,8 +1,12 @@
 ## maio Android Mediation Adapter Changelog
 
-#### Next Version
+#### Version 2.0.8.2
 - Filter age-restricted requests for maio SDK because it cannot receive age-restricted signals.
 - Added property to build the adapter with GMA Next-Gen SDK dependency.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.1.0.
+- Maio SDK version 2.0.8.2.
 
 #### Version 2.0.8.1
 - Added Banner ad format support