+++ ThirdPartyAdapters/pangle/CHANGELOG.md
@@ -1,6 +1,18 @@
 ## Pangle Android Mediation Adapter Changelog
 
-#### Version 7.9.1.3.0 (In progress)
+#### Version 8.0.0.4.0
+- Verified compatibility with Pangle SDK version 8.0.0.4.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.1.0.
+- Pangle SDK version 8.0.0.4.
+
+#### Version 7.9.1.3.0
+- Verified compatibility with Pangle SDK version 7.9.1.3.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.1.0.
+- Pangle SDK version 7.9.1.3.
 
 #### Version 7.9.1.2.0
 - Added property to build the adapter with GMA Next-Gen SDK dependency.
@@ -8,6 +20,7 @@
 
 Built and tested with:
 - Google Mobile Ads SDK version 25.1.0.
+- Google Mobile Ads Next-Gen SDK version 0.25.0-beta01.
 - Pangle SDK version 7.9.1.2.
 
 #### Version 7.9.1.1.0